+++ mailscanner/viewpart.php
@@ -42,29 +42,31 @@ require 'login.function.php';
 ini_set("memory_limit", MEMORY_LIMIT);
 
 if (!isset($_GET['id'])) {
-    die("No input Message ID");
+    die('No input Message ID');
 } else {
+    $message_id = sanitizeInput($_GET['id']);
     // See if message is local
     dbconn(); // required db link for mysql_real_escape_string
-    if (!($host = @mysql_result(
+    $message_data = mysql_fetch_object(
         dbquery(
-            "SELECT hostname FROM maillog WHERE id='" . mysql_real_escape_string(
-                $_GET['id']
-            ) . "' AND " . $_SESSION["global_filter"] . ""
-        ),
-        0
-    ))
-    ) {
-        die("Message '" . $_GET['id'] . "' not found\n");
+            "SELECT hostname, DATE_FORMAT(date,'%Y%m%d') AS date FROM maillog WHERE id='" .
+            mysql_real_escape_string($message_id) . "' AND "
+            . $_SESSION["global_filter"]
+        )
+    );
+
+    if (!$message_data) {
+        die("Message '" . $message_id . "' not found\n");
     }
-    if (!is_local($host) || RPC_ONLY) {
+
+    if (!is_local($message_data->hostname) || RPC_ONLY) {
         // Host is remote - use XML-RPC
         //$client = new xmlrpc_client(constant('RPC_RELATIVE_PATH').'/rpcserver.php', $host, 80);
-        $input = new xmlrpcval($_GET['id']);
+        $input = new xmlrpcval($message_id);
         $parameters = array($input);
         $msg = new xmlrpcmsg('return_quarantined_file', $parameters);
         //$rsp = $client->send($msg);
-        $rsp = xmlrpc_wrapper($host, $msg);
+        $rsp = xmlrpc_wrapper($message_data->hostname, $msg);
         if ($rsp->faultcode() == 0) {
             $response = php_xmlrpc_decode($rsp->value());
         } else {
@@ -72,42 +74,28 @@ if (!isset($_GET['id'])) {
         }
         $file = base64_decode($response);
     } else {
-        $date = @mysql_result(
-            dbquery(
-                "SELECT DATE_FORMAT(date,'%Y%m%d') FROM maillog where id='" . mysql_real_escape_string(
-                    $_GET['id']
-                ) . "' AND " . $_SESSION["global_filter"] . ""
-            ),
-            0
-        );
-        $qdir = get_conf_var('QuarantineDir');
+        //build filename path
+        $quarantine_dir = get_conf_var('QuarantineDir');
+        $filename = '';
         switch (true) {
-            case (file_exists($qdir . '/' . $date . '/nonspam/' . $_GET['id'])):
-                $_GET['filename'] = $date . '/nonspam/' . $_GET['id'];
+            case (file_exists($quarantine_dir . '/' . $message_data->date . '/nonspam/' . $message_id)):
+                $filename = $message_data->date . '/nonspam/' . $message_id;
                 break;
-            case (file_exists($qdir . '/' . $date . '/spam/' . $_GET['id'])):
-                $_GET['filename'] = $date . '/spam/' . $_GET['id'];
+            case (file_exists($quarantine_dir . '/' . $message_data->date . '/spam/' . $message_id)):
+                $filename = $message_data->date . '/spam/' . $message_id;
                 break;
-            case (file_exists($qdir . '/' . $date . '/mcp/' . $_GET['id'])):
-                $_GET['filename'] = $date . '/mcp/' . $_GET['id'];
+            case (file_exists($quarantine_dir . '/' . $message_data->date . '/mcp/' . $message_id)):
+                $filename = $message_data->date . '/mcp/' . $message_id;
                 break;
-            case (file_exists($qdir . '/' . $date . '/' . $_GET['id'] . '/message')):
-                $_GET['filename'] = $date . '/' . $_GET['id'] . '/message';
+            case (file_exists($quarantine_dir . '/' . $message_data->date . '/' . $message_id . '/message')):
+                $filename = $message_data->date . '/' . $message_id . '/message';
                 break;
         }
 
-        // File is local
-        if (!isset($_GET['filename'])) {
-            die("No input filename");
-        } else {
-            // SECURITY - strip off any potential nasties
-            $_GET['filename'] = preg_replace('[\.\/|\.\.\/]', '', $_GET['filename']);
-            $filename = get_conf_var('QuarantineDir') . "/" . $_GET['filename'];
-            if (!@file_exists($filename)) {
-                die("Error: file not found\n");
-            }
-            $file = file_get_contents($filename);
+        if (!@file_exists($quarantine_dir . '/' . $filename)) {
+            die("Error: file not found\n");
         }
+        $file = file_get_contents($quarantine_dir . '/' . $filename);
     }
 }
 
@@ -159,7 +147,7 @@ function decode_structure($structure)
                 $structure->body = utf8_encode($structure->body);
             }
             if (STRIP_HTML) {
-                $structure->body = str_replace('<!DOCTYPE','<DOCTYPE', $structure->body);
+                $structure->body = str_replace('<!DOCTYPE', '<DOCTYPE', $structure->body);
                 echo strip_tags($structure->body, ALLOWED_TAGS);
             } else {
                 echo $structure->body;